+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-- Added Highlights Outline to the Blocks
+- Fixed effects doesnt work properly on armor